requirements.txt
asgiref==3.6.0
Django==4.2
psycopg2-binary==2.9.6
sqlparse==0.4.3
djangorestframework==3.13.1
django-filter==21.1